pyproject.toml
[project]
name = "dsad-tech-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "joblib>=1.4.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
]
